requirements.txt
nonebot2[fastapi]
nonebot_adapter_mirai2
sqlalchemy
pydantic
pymysql
cryptography